requirements.txt
annotated-types==0.7.0
black==25.12.0
blessed==1.25.0
build==1.4.0
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.1
distlib==0.4.0
editor==1.6.6
filelock==3.20.3
flake8==7.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
inquirer==3.4.1
isort==7.0.0
librt==0.7.7
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.1
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
readchar==4.2.1
requests==2.32.5
rich==14.2.0
rsa==4.9.1
runs==1.2.2
shellingham==1.5.4
tqdm==4.67.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.36.1
wcwidth==0.2.14
xmod==1.8.1
python-dotenv==1.2.1
